Moved component file
