+++ __tests__/core/ComponentMap.test.ts
@@ -1,4 +1,4 @@
-import { Component } from '../../src/engine/components/Component';
+import { Component } from '../../src/engine/core/Component';
 import ComponentMap from '../../src/engine/core/ComponentMap';
 import { Entity } from '../../src/engine/core/EntityStore';
