+++ src/engine/components/Position.ts
@@ -1,4 +1,4 @@
-import { Component } from "./Component";
+import { Component } from "../core/Component";
 
 export class Position implements Component {
   constructor(public x: number, public y: number) {}
