+++ src/engine/components/Rotation.ts
@@ -1,4 +1,4 @@
-import { Component } from "./Component";
+import { Component } from "../core/Component";
 
 export class Rotation implements Component {
   constructor(public rotation: number) {}
